added Detector green

+++ src/Motor.cpp
@@ -1,8 +1,9 @@
 #include "Motor.h"
 #include "Arduino.h"
 
-Motor::Motor(int mounting_deg, uint8_t pin_fw, uint8_t pin_bw)
+Motor::Motor(int mounting_deg, int pin_fw, int pin_bw)
 {
+    // pinMode(Output) wenn nicht in main setup
     mounting_offset = mounting_deg;
     p_fw = pin_fw;
     p_bw = pin_bw;
