+++ src/Motor.h
@@ -14,11 +14,11 @@ class Motor
 {
 private:
     int mounting_offset;
-    uint8_t p_fw;
-    uint8_t p_bw;
+    int p_fw;
+    int p_bw;
 
 public:
-    Motor(int mounting_deg, uint8_t pin_fw, uint8_t pin_bw);
+    Motor(int mounting_deg, int pin_fw, int pin_bw); // pinMode(Output) wenn nicht in main setup
     ~Motor();
     void drive_direction(int degree);
     void stop();
